Modify JSON parsing to not choke to death on profile pictures. ...
Facbook's API returns a raw JPEG via get_connections(fb_id, "picture"),
so we have to deal with that here and fail JSON parsing gracefully.

This could be implemented to force-wrap the object in a JSON response so
that there is no confusion regarding return values and we always get a
dictionary back as response.